backend/pyproject.toml
[tool.poetry]
name = "todo-chatbot-backend"
version = "0.1.0"
description = "Backend for the Todo Chatbot application"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.5.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
uuid = "^1.30"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.main:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | downloads
  | node_modules
  | venv
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
known_third_party = ["fastapi", "uvicorn", "sqlalchemy", "alembic", "pydantic", "pytest", "httpx", "passlib", "python_jose", "dotenv"]
known_first_party = ["src"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    "node_modules",
    ".venv",
    "alembic"
]